requirements.txt
asgiref==3.7.2
astroid==3.0.1
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
playwright==1.39.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pyee==11.0.1
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
